+++ .eslintrc.js
@@ -1,0 +1,27 @@
+module.exports = {
+  env: {
+    commonjs: true,
+    es2021: true,
+    node: true,
+  },
+  extends: ['airbnb-base', 'prettier'],
+  plugins: ['prettier'],
+  parserOptions: {
+    ecmaVersion: 12,
+  },
+  rules: {
+    'prettier/prettier': 'error',
+    indent: ['error', 2],
+    'linebreak-style': ['error', 'unix'],
+  },
+  overrides: [
+    {
+      files: ['**/*.test.js'],
+      env: { jest: true },
+      plugins: ['jest'],
+      // eslint-disable-next-line global-require, import/no-extraneous-dependencies
+      ...require('eslint-plugin-jest').configs.recommended,
+    },
+  ],
+  ignorePatterns: ['**/node_modules/**', '**/coverage/**', '**/doc/**'],
+};